requirements.txt
# Core dependencies
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
numpy==1.24.3
pandas==2.0.3

# NLP and AI
transformers==4.35.2
sentence-transformers==2.2.2
openai==1.3.0
langchain==0.0.350
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1

# Web interfaces
streamlit==1.29.0
gradio==4.8.0
fastapi==0.104.1
uvicorn==0.24.0

# Communication
python-telegram-bot==20.7
twilio==8.10.0

# Database and caching
redis==5.0.1
pymongo==4.6.0

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
tqdm==4.66.1
reportlab==4.0.7
python-multipart==0.0.6
aiohttp==3.9.1